Drag'n'Drop Cover-Image between AmazonMusic App and WinMusik editor

+++ build_mingw64.sh
@@ -25,6 +25,9 @@ fi
 cat include/version.h
 
 mingw32-make -j8 release
+if [ ?$ -ne 0 ] ; then
+   exit 1
+fi
 rm -rf release/deploy
 windeployqt.exe --dir release/deploy release/WinMusik.exe
